pyproject.toml
[project]
name = "embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "einops>=0.8.0",
    "matplotlib>=3.10.0",
    "opentsne>=1.0.2",
    "pyqt5>=5.15.11",
    "pyside6>=6.8.1.1",
    "python-fasthtml>=0.12.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.4.0",
    "torch>=2.5.1",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
]
